app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
click==8.1.7
distlib==0.3.9
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
MarkupSafe==3.0.2
numpy==2.1.3
openai==1.55.0
packaging==24.2
platformdirs==4.3.6
pydantic==2.10.0
pydantic_core==2.27.0
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
virtualenv==20.27.1
Werkzeug==3.1.3
